FLOE-255: Implemented a new textSize panel
Implemented a new text size panel based on the first discovery tool designs. This differs from the one in Infusion in that it doesn't use a textFieldSlider and has a meter indicator instead. The integration doesn't work yet, as there are some issues with the markup that may be addressed in work on FLOE-252.