FLOE-299: refactored visibility for assistance
Using jQuery's toggle to hide/show with display:none instead of using visibility:hidden. This will keep the input field close to the text, but will cause it to shift as the panel's states change. This has been determined to be better than leaving a large initial gap between the text and the input.