Foundation Tooltips and Initial Keyboard Navigation.
I was changing some things in the firstDiscoveryEditor to support the
't' button command for the fdt that will toggle whether or not the
preview is focused or not. I was having some issues with getting the
infusion to recognize the functions. On my to do list. I also did some
testing on incorporating preview built foundation tooltips that will
hopefully be self voiced here at some point.