Bug Fixes I
Changed the following:
- The French language had a separate width for the the voice toggle in
the style.css. I updated that to reflect previous changes. (made it
wider).
- I added a commands key to the panels.js file under the welcome panel.
This will allow me to print off the help message without voicing it.
This eliminates the double help message voicing.
- The entire tool uses outside borders to signify focus or hover. The
issue with the voice toggle was that it was using an internal border
that changed the padding. I removed that and used the focus formatting
for the hover on the voicing button. This eliminates the icon movement
issue.
- I noticed another issue pertaining to the voice toggle button getting
focus and having a strange blue background color. I removed this from
the css as well.